+++ test/components/monthview/monthview.e2e-spec.js
@@ -206,15 +206,6 @@ describe('MonthView disable day tests', () => {
     expect(await element.all(by.css('.monthview-table td')).count()).toEqual(42);
     expect(await element.all(by.css('.monthview-table td.is-disabled')).count()).toEqual(12);
   });
-
-  if (utils.isChrome() && utils.isCI()) {
-    it('Should not visual regress', async () => {
-      const containerEl = await element(by.className('container'));
-      await browser.driver.sleep(config.sleepLonger);
-
-      expect(await browser.protractorImageComparison.checkElement(containerEl, 'monthview-weekends')).toEqual(0);
-    });
-  }
 });
 
 describe('MonthView disable month selection tests', () => {